server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
redis==4.5.5
pika==1.3.2
psycopg2-binary